Bugfix: create 'Habilidade' and 'Beneficio' at 'Vaga' views

+++ Controllers/EmpresasController.cs
@@ -2,6 +2,7 @@
 using EmpregosOnLine.Models.ViewModels;
 using EmpregosOnLine.Services;
 using Microsoft.AspNetCore.Mvc;
+using System.Diagnostics;
 
 namespace EmpregosOnLine.Controllers
 {
@@ -24,14 +25,14 @@ namespace EmpregosOnLine.Controllers
         {
             if (id == null)
             {
-                return NotFound();
+                return RedirectToAction(nameof(Error), new { message = "Ocorreu um erro durante a requisição das informações." });
             }
             
             var empresa = await _empresasService.GetEmpresaAsync(id.Value);
 
             if (empresa == null)
             {
-                return NotFound();
+                return RedirectToAction(nameof(Error), new { message = "Empresa não encontrada." });
             }
 
             return View(empresa);
@@ -55,14 +56,13 @@ namespace EmpregosOnLine.Controllers
             return View(empresaEnderecoViewModel);
         }
 
-        // GET: EmpresasController/Edit/5
         public async Task<ActionResult> Edit(Guid id)
         {
             var empresa = await _empresasService.GetEmpresaAsync(id);
             
             if (empresa == null)
             {
-                return NotFound();
+                return RedirectToAction(nameof(Error), new { message = "Ocorreu um erro durante a requisição das informações." });
             }
             
             var viewModel = new EmpresaEnderecoViewModel { Id = empresa.Id, Empresa = empresa, Endereco = empresa.Endereco };
@@ -87,7 +87,7 @@ namespace EmpregosOnLine.Controllers
             var empresa = await _empresasService.GetEmpresaAsync(id);
             if (empresa == null)
             {
-                return NotFound();
+                return RedirectToAction(nameof(Error), new { message = "Ocorreu um erro durante a requisição das informações." });
             }
             return View(empresa);
         }
@@ -98,8 +98,25 @@ namespace EmpregosOnLine.Controllers
         public async Task<ActionResult> DeleteConfirmed(Guid id)
         {
             var empresa = await _empresasService.GetEmpresaAsync(id);
+
+            if (empresa.Vagas.Count() > 0)
+            {
+                return RedirectToAction(nameof(Error), new { message = "Não é possível deletar empresas que possuem uma ou mais vagas associadas." });
+            }
+
             await _empresasService.DeleteEmpresaAsync(empresa);
             return RedirectToAction(nameof(Index));
+        }
+
+        public IActionResult Error(string message) 
+        {
+            var viewModel = new ErrorViewModel
+            {
+                Message = message,
+                RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier
+            };
+
+            return View(viewModel);
         }
     }
 }
